Add global Mage\Framework\View\Layout preference
We were running into an issue where the DepersonalizePlugin was clearing
the checkout session data when rendering the cart page, resulting in an
eventual extension trying to load a non-existent quote.

I've traced the issue to a layout that was being instantiated in
Magento\Framework\Pricing\Render having the cacheable flag set to true
(it's false in vanilla M2 in that instance) due to the parent layout
being cacheable. However, I couldn't figure out why this was happening.
My theory is that by replacing the layout model used in the context of
our own page model, we've caused a new layout to be instantiated which
does not inherit the cacheable flag from core Magento for some reason.
Replacing the Layout model as a global preference fixes the issue.

Fixes #5