Fix duplicate content on some SW routes (Fixes #8)
The issue was happening because the layout was not yet built when we
removed the "root" output element to replace it with "columns". This
meant that there was nothing to remove (which throws no errors) and when
the layout gets built at a later time, the "root" element gets added to
the output after "columns", causing the content to be rendered twice.